Create apps.go to refactor Applications related common methods
